refactor(epicenter): rename src/static/ to src/workspace/
The 'static' name only made sense as contrast to the now-deleted Dynamic
API. Rename to 'workspace' which accurately describes what this folder
contains — the workspace implementation.

- git mv src/static → src/workspace (25 files)
- Update all internal imports within the package
- Update package.json ./static export to point to new path (backward compat)
- Migrate epicenter app imports from @epicenter/hq/static to @epicenter/hq
- Clean up stale 'Static API' naming in types.ts JSDoc and README.md
- Update deprecation spec with rename completion
