+++ apps/epicenter/src/lib/templates/entries.ts
@@ -9,7 +9,7 @@
  * - tags: additional tagging (string)
  */
 
-import { defineTable, defineWorkspace } from '@epicenter/hq/static';
+import { defineTable, defineWorkspace } from '@epicenter/hq';
 import { type } from 'arktype';
 
 const entries = defineTable(
