+++ apps/epicenter/src/lib/templates/whispering.ts
@@ -6,7 +6,7 @@
  * Epicenter workspace.
  */
 
-import { defineTable, defineWorkspace } from '@epicenter/hq/static';
+import { defineTable, defineWorkspace } from '@epicenter/hq';
 import { type } from 'arktype';
 
 const recordings = defineTable(
