+++ packages/epicenter/scripts/stress-test-static.ts
@@ -11,7 +11,7 @@
 import { unlinkSync } from 'node:fs';
 import { type } from 'arktype';
 import * as Y from 'yjs';
-import { createTables, defineTable } from '../src/static/index.js';
+import { createTables, defineTable } from '../src/workspace/index.js';
 
 // ═══════════════════════════════════════════════════════════════════════════════
 // Config — tweak these
